RAGTask/pyproject.toml
[project]
name = "rag-task"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "groq>=0.32.0",
    "langchain>=0.3.27",
    "langfuse>=3.6.2",
    "litellm>=1.78.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
